move bitmap to bitmap.go and move ocr to img.go
