+++ base/file_io_c.h
@@ -1,6 +1,6 @@
 #include "file_io.h"
 // #include "os.h"
-// #include "bmp_io_c.h"
+#include "bmp_io_c.h"
 #include "png_io_c.h"
 #include <stdio.h> /* For fputs() */
 #include <string.h> /* For strcmp() */
